Eliminate boot files
Here we resolve the import loop between Encoding and FlatTerm by moving their
types to dedicated modules, allowing FlatTerm to import Encoding.Types, enabling
it to define toFlatTerm. Encoding can then import FlatTerm.toFlatTerm for use
in the (now orphaned) Show Encoding instance. Despite the orphan, I think this
is the best we can do since FlatTerm and Encoding can't be defined in the same
source file due to overlapping constructor names.

Fixes #104.
