--- src/Serialise/Cborg/Encoding.hs-boot
@@ -1,5 +1,0 @@
-module Serialise.Cborg.Encoding where
-
-newtype Encoding = Encoding (Tokens -> Tokens)
-
-data Tokens
