--- src/Serialise/Cborg/FlatTerm.hs-boot
@@ -1,10 +1,0 @@
-module Serialise.Cborg.FlatTerm where
-
-import {-# SOURCE #-} Serialise.Cborg.Encoding
-
-type FlatTerm = [TermToken]
-
-data TermToken
-instance Show TermToken
-
-toFlatTerm :: Encoding -> FlatTerm
